finalise the blocks
